# server/openapi.yaml
openapi: 3.1.0
...
info:
  title: ANZH Beauty API
  version: '1.0.0'
  description: |
    Бэкенд мини-приложения ANZH Cosmetology.

    **Личность.** Все запросы, кроме витрины и здоровья, требуют
    сессии. Сессия выдаётся в обмен на `initData` из Telegram, подпись
    которого проверяется на сервере (HMAC + срок жизни 1 час). Роль
    берётся с сервера по `telegram_id`, а не из запроса.

    **Деньги.** Целое в минорных единицах плюс код валюты: `45000` и
    `GEL` означают 450.00 лари. Telegram Stars — отдельное поле
    `stars_amount`, это не валюта ISO-4217.

    **Время.** Всё в UTC, ISO-8601. Часовой пояс салона —
    `Asia/Tbilisi` — параметр конфигурации, а не свойство данных.

    **Итог визита считает сервер.** Клиент присылает состав, а не сумму.
servers:
  - url: '{base}/api/v1'
    description: Базовый адрес берётся из переменной окружения клиента
    variables:
      base:
        default: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /team:
    get:
      tags: [Команда]
      summary: Состав студии
      description: Открыт без сессии. Ни PIN, ни его хэш не отдаются никогда.
      security: []
      responses:
        '200':
          description: Команда
          content:
            application/json:
              schema:
                type: object
                properties:
                  team:
                    type: array
                    items: { $ref: '#/components/schemas/StaffMember' }
    post:
      tags: [Команда]
      summary: Добавить или вернуть сотрудника
      description: Только владелица. Повторный вызов с тем же `telegram_id` обновляет запись.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [telegram_id, name]
              properties:
                telegram_id: { type: integer, format: int64 }
                name: { type: string }
                title: { type: [string, 'null'] }
                role: { type: string, enum: [staff, owner], default: staff }
      responses:
        '201':
          description: Заведён
          content:
            application/json:
              schema:
                type: object
                properties:
                  staff: { $ref: '#/components/schemas/StaffMember' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Invalid' }
  /team/{telegramId}:
    parameters:
      - { name: telegramId, in: path, required: true, schema: { type: integer, format: int64 } }
    delete:
      tags: [Команда]
      summary: Снять доступ
      description: |
        Мягко: записи этого мастера остаются в истории и в выручке.
        Снять доступ у себя нельзя — вернуть его через приложение
        было бы уже нечем.
      responses:
        '200':
          description: Доступ снят
          content:
            application/json:
              schema:
                type: object
                properties: { deactivated: { type: integer, format: int64 } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: '`self_removal`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /team/{telegramId}/services:
    parameters:
      - { name: telegramId, in: path, required: true, schema: { type: integer, format: int64 } }
    put:
      tags: [Команда]
      summary: Что ведёт мастер
      description: Полная замена набора — снятие процедуры работает тем же запросом, что добавление.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slugs]
              properties:
                slugs: { type: array, items: { type: string } }
      responses:
        '200':
          description: Набор заменён
          content:
            application/json:
              schema:
                type: object
                properties:
                  staff_id: { type: integer, format: int64 }
                  services: { type: array, items: { type: string } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422':
          description: '`bad_slugs` · `unknown_service`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /staff/me:
    get:
      tags: [Команда]
      summary: Кабинет — кто я как сотрудник
      responses:
        '200':
          description: Сотрудник
          content:
            application/json:
              schema:
                type: object
                properties:
                  staff:
                    allOf:
                      - $ref: '#/components/schemas/StaffMember'
                      - type: object
                        properties: { pin_set: { type: boolean } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: '`not_staff`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /staff/pin:
    post:
      tags: [Команда]
      summary: Подтвердить PIN
      description: |
        **Это не вход.** Вход уже состоялся через `initData`, роль взята
        из базы. PIN — второй фактор: подтверждает, что за телефоном
        тот же человек. Кто не мастер, тот им не станет, введя верный
        код.

        Ограничение частоты: 5 попыток за 5 минут.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pin]
              properties: { pin: { type: string } }
      responses:
        '200':
          description: Подтверждён
          content:
            application/json:
              schema:
                type: object
                properties: { confirmed: { type: boolean, const: true } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: '`forbidden` — не сотрудник · `bad_pin` — код не подошёл'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: '`bad_pin` · `pin_not_set`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/TooMany' }
  /staff/pin/set:
    post:
      tags: [Команда]
      summary: Задать или сменить свой PIN
      description: |
        Хранится только хэш (scrypt с солью). Демо-коды из спецификации
        (`2024`, `1111`, `2222`) ходили по переписке и отклоняются.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pin]
              properties: { pin: { type: string, pattern: '^\d{4,8}$' } }
      responses:
        '200':
          description: Установлен
          content:
            application/json:
              schema:
                type: object
                properties: { updated: { type: boolean, const: true } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422':
          description: '`bad_pin` · `pin_leaked` — код известен из переписки'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <токен>`. Полезная нагрузка ровно такая:
        `{ telegram_id, app_source: "beauty", role }`. Поле `app_source`
        сверяется при каждом запросе — токен другого приложения это не
        открывает.
...
  responses:
    Unauthorized:
      description: Нет сессии
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: no_session, message: Нужно войти заново } }
...
    Forbidden:
      description: Роли или прав на объект не хватает
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Объекта нет
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Invalid:
      description: Не проходит проверку
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooMany:
      description: Слишком часто. Заголовок `Retry-After` — через сколько секунд можно снова
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      description: |
        Единая форма на все отказы. Ни трассировки стека, ни текста
        запроса к базе: по ним читается устройство сервера, а в тексте
        запроса могут оказаться данные из анкеты.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: slot_taken }
            message: { type: string, example: Время уже занято }
...
    StaffMember:
      type: object
      properties:
        telegram_id: { type: integer, format: int64 }
        name: { type: string }
        title: { type: [string, 'null'] }
        role: { type: string, enum: [staff, owner] }
        photo_url: { type: [string, 'null'] }
        services: { type: array, items: { type: string } }
...